refactor: move yajl_tree to tree
