+++ crates/yajl-clib/src/yajl_tree.rs
@@ -4,7 +4,7 @@
 
 use ::libc;
 
-pub use yajl::yajl_tree::{yajl_type, yajl_val};
+pub use yajl::tree::{yajl_type, yajl_val};
 
 #[no_mangle]
 pub unsafe extern "C" fn yajl_tree_parse(
@@ -12,7 +12,7 @@ pub unsafe extern "C" fn yajl_tree_parse(
     mut error_buffer: *mut libc::c_char,
     mut error_buffer_size: usize,
 ) -> yajl_val {
-    yajl::yajl_tree::yajl_tree_parse(input, error_buffer, error_buffer_size)
+    yajl::tree::yajl_tree_parse(input, error_buffer, error_buffer_size)
 }
 
 #[no_mangle]
@@ -21,10 +21,10 @@ pub unsafe extern "C" fn yajl_tree_get(
     mut path: *mut *const libc::c_char,
     mut type_0: yajl_type,
 ) -> yajl_val {
-    yajl::yajl_tree::yajl_tree_get(n, path, type_0)
+    yajl::tree::yajl_tree_get(n, path, type_0)
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn yajl_tree_free(mut v: yajl_val) {
-    yajl::yajl_tree::yajl_tree_free(v)
+    yajl::tree::yajl_tree_free(v)
 }
